03-ReactBasics: 19. Card Component
CardList is the one who responsible to render the cards. So, it is a stateless component, it takes the props from the app and just render the card items.
We destruct the property monsters from props.
We create another component Card for each Card item in the CardList.

Card is also a stateless component (here, we didn't use prop-types).
Card is a div with class card-container, and it contains img, h2, and p. The src of the image if from robohash.org api.